Attempt to fix GTK hid zooming/panning/flipping

+++ src/hid/common/trackball.h
@@ -66,7 +66,7 @@ add_quats(float *q1, float *q2, float *dest);
  * given quaternion.
  */
 void
-build_rotmatrix(float m[4][4], float q[4]);
+build_rotmatrix(double m[4][4], float q[4]);
 
 /*
  * This function computes a quaternion based on an axis (defined by
